Move services to application domain

+++ src/Application/Service/DeleteProductService.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Bdd\Domain\Service;
+namespace Bdd\Application\Service;
 
+use Bdd\Application\Request\Error\ResourceNotFoundException;
 use Bdd\Domain\Repository\ProductRepository;
 
 class DeleteProductService
@@ -17,6 +18,10 @@ class DeleteProductService
     public function delete(string $id): void
     {
         $product = $this->productRepository->find($id);
+
+        if (!$product) {
+            throw new ResourceNotFoundException('Product not found');
+        }
 
         $this->productRepository->remove($product);
     }
